pyproject.toml
[tool.poetry]
name = "documento"
version = "0.1.0"
description = "A small Document Management System (DMS)"
authors = ["Jonathan Weth <dev@jonathanweth.de>"]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.6"
django-rest-framework = "^0.1.0"
django = "^3.1.1"
django-rest-knox = "^4.1.0"
django-webpack-loader = "^0.7.0"
django-cleanup = "^5.1.0"
blabel = "^0.1.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
